LPS-47465 add ServiceReferenceMapper interface
Users of ServiceTrackerMap must provice an implementation of
ServiceReferenceMapper in order for the Map to obtain the keys under
which store the services related to a ServiceReference.

The way to provide the keys is invoking Emitter.emit(K key) with each
key under which the service must be stored.

I chose this interface because:

* It allows to provide 0, 1 or N keys for the same object
* It does not force the client of the API to reserve any memory

It is also a common pattern when some logic needs to provide keys for an
object, see:

http://hadooptutorial.wikispaces.com/MapReduce <-- look for
Emit(intermediateKey, intermediateValue);

http://wiki.apache.org/couchdb/Introduction_to_CouchDB_views#Map_Functions

http://docs.cloudant.com/guides/mapreduce.html#map

http://docs.mongodb.org/manual/reference/command/mapReduce/#mapreduce-map-cmd

In our case emit signature only has the key since this is an implementation
bounded to services and ServiceReferences from the Registry.