LPS-61767 Make them read-only for now
These properties are used immediately after the plugin is applied, and if someone changes their values later in the extension object, nothing happens. We could let users set these values through the extension object, but it would require a deeper refactoring. So it's better to keep them read-only for now.