LPS-63462 Extender to publish the ResourceBundleLoaders based on capabilities
This extender will register a ResourceBundleLoader for those bundles
declaring a capability for a bundle. If a bundle provides a capability
with an `aggregate` attribute and a comma separated list of OSGi filters
it will publish an AggregateResourceBundleLoader that tracks those filters.
This way a bundle can declare that it's keys are a composition of some other
bundles' keys:

```Provide-Capability:
liferay.resource.bundle;aggregate="(bundle.symbolic.name=modulo1),(bundle.symbolic.name=modulo2)";myAttribute=myValue```

This AggregateResourceBundleLoader will be registered as the bundle's
default ResourceBundleLoader