LPS-64277 Always execute UpgradePortletId
We cannot add KB to the UpgradeModules upgrade, as it may not be
available in all deployments. If we did so, we would create an unused
entry in the Release_ table; it does no harm to always execute the
UpgradePortletId task, as it will only modify the DB if it needs to, and
performance-wise it's not very demanding.